reroute stuff, add button group
