+++ src/lib/components/DataInput/AbstractDataInput.svelte
@@ -1,8 +1,9 @@
 <script>
-	import { Heading } from 'flowbite-svelte';
+	import { Button, ButtonGroup, Heading } from 'flowbite-svelte';
 	export let props = [];
 	export let heading = null;
 	export let description = null;
+	export let buttons = null;
 </script>
 
 <div class="space-y-6rtl:space-x-reverse items-center space-x-4">
@@ -33,5 +34,19 @@
 			{/if}
 			<svelte:component this={prop.itemComponent} {...prop.componentProps} />
 		{/each}
+
+		{#if buttons}
+			<ButtonGroup>
+				{#each buttons as bprops}
+					<Button
+						outline
+						pill={true}
+						href={bprops.href}
+						class="text-gray-500 hover:text-primary-800 group-hover:text-primary-800 dark:text-gray-400 dark:hover:text-primary-500 dark:group-hover:text-primary-500"
+						>{bprops.label}</Button
+					>
+				{/each}
+			</ButtonGroup>
+		{/if}
 	</form>
 </div>
