+++ src/lib/components/DataInput/AbstractRegistrationForm.svelte
@@ -1,10 +1,9 @@
 <script>
-	import { Button, Card, Heading } from 'flowbite-svelte';
+	import { Card, Heading } from 'flowbite-svelte';
 	import AbstractDataInput from './AbstractDataInput.svelte';
 	export let props = [];
 	export let heading = null;
-	export let buttonLabel = null;
-	export let href = '';
+	export let buttons = null;
 </script>
 
 <div class="m-1 items-center justify-center pb-6">
@@ -16,11 +15,6 @@
 				>{heading}</Heading
 			>
 		{/if}
-		<!-- <Label>{label}</Label>
-			<Input {type} {label} {placeholder} /> -->
-		<AbstractDataInput {props} />
-		{#if buttonLabel}
-			<Button type="submit" {href} class="w-full">{buttonLabel}</Button>
-		{/if}
+		<AbstractDataInput {props} {buttons} />
 	</Card>
 </div>
