+++ src/routes/userLand/userDataValidation/+page.svelte
@@ -4,29 +4,16 @@
 	import AbstractRegistrationForm from '$lib/components/DataInput/AbstractRegistrationForm.svelte';
 	import { Input } from 'flowbite-svelte';
 
-	// this stuff here will become backend calls in the end because that is where the data this page will be filled with
-	// will come from. Hence, they are not put into a separate library or anything
-	function intervalRange(size, startAt = 0, step = 1) {
-		return [...Array(size).keys()].map(
-			(i) => String(i * step + startAt) + '-' + String((i + 1) * step + startAt)
-		);
-	}
-
-	function numericalRange(size, startAt = 0, step = 1) {
+	function range(size, startAt = 0, step = 1) {
 		return [...Array(size).keys()].map((i) => i * step + startAt);
 	}
 
-	export let role = 'Admin';
-
-	const heading = 'Benutzerdaten eingeben';
-	const description =
-		'Bitte vervollständigen sie ihre Daten um die Registrierung abzuschließen. Wählen sie dazu das jeweils zutreffende aus der jeweiligen Liste aus';
-	const user_props = [
+	const props = [
 		{
 			itemComponent: AbstractDropdownItem,
 			componentProps: {
 				name: 'Geburtsjahr',
-				items: numericalRange(100, 1920),
+				items: range(100, 1920),
 				about: 'Wählen sie ihr Geburtsjahr aus',
 				selected: 1988
 			}
@@ -62,20 +49,20 @@
 			itemComponent: AbstractDropdownItem,
 			componentProps: {
 				name: 'Arbeitszeit/Woche',
-				items: intervalRange(13, 0, 5),
+				items: range(13, 0, 5),
 				about:
 					'Wählen sie ihre Arbeitszeit pro Woche aus. Wählen sie die Zahl, die dem tatsächlichen Wert am nächsten kommt.',
-				selected: '35-40'
+				selected: 40
 			}
 		},
 		{
 			itemComponent: AbstractDropdownItem,
 			componentProps: {
 				name: 'Familieneinkommen/Jahr',
-				items: intervalRange(23, 0, 5000),
+				items: range(23, 0, 5000),
 				about:
 					'Wählen sie ihre Jahreseinkommen aus. Wählen sie die Zahl, die dem tatsächlichen Wert am nächsten kommt.',
-				selected: '50000-55000'
+				selected: 50000
 			}
 		},
 
@@ -85,56 +72,23 @@
 				name: 'Beruf',
 				type: 'text',
 				about: 'Geben sie ihren Beruf an',
-				placeholder: 'Geben sie ihren Beruf an'
+				placeholder: 'Geben sie ihren Beruf an',
+				value: 'software engineer'
 			}
 		}
 	];
+	const heading = 'Überprüfen sie ihre Daten';
+	const description = 'Bitte überprüfen sie ihre Daten und korrigieren sie diese gegebenenfalls.';
 
-	let props = user_props;
-
-	$: if (role === 'Forscher/in') {
-		props = user_props;
-		props.push({
-			itemComponent: AbstractDropdownItem,
-			componentProps: {
-				name: 'Forschungsbereich',
-				items: ['Psychologie', 'Medizin', 'Physik'],
-				about: 'Wählen sie ihren Forschungsbereich aus',
-				selected: 'Psychologie'
-			}
-		});
-
-		props.push({
-			itemComponent: Input,
-			componentProps: {
-				name: 'Institution',
-				type: 'text',
-				about: 'Geben sie an an welcher Forschungsinstitution sie tätig sind',
-				placeholder: 'Forschungsinstitution'
-			}
-		});
-	}
-
-	$: if (role == 'Admin') {
-		props = user_props;
-		props.push({
-			itemComponent: Input,
-			componentProps: {
-				name: 'Institution',
-				type: 'text',
-				about: 'Geben sie ihr Admin Token ein',
-				placeholder: 'Admin Token'
-			}
-		});
-	}
+	const buttons = [{ label: 'Abschicken', href: '/' }];
 </script>
 
 <div class="m-1 w-full max-w-md items-center justify-center">
 	<AbstractContent
 		infopage={'/info'}
-		lastpage={'/registerUser'}
-		nextpage={'/registerUser/userDataValidation'}
+		lastpage={'/registerUser/userDataInput'}
+		nextpage={'/registerChild'}
 	>
-		<AbstractRegistrationForm {props} {heading} {description} />
+		<AbstractRegistrationForm {props} {heading} {description} {buttons} />
 	</AbstractContent>
 </div>
